+++ CHANGES.md
@@ -139,3 +139,14 @@
 - Added possibilities to align all types of input- and reference-geometries ((multi-)point,line and polygons) [#191]
 - Added parameter 'preserve_topology' to keep relations between imput-geometries [#90]
 - Bugfix on _evaluate-function [#185]
+
+# 0.11.0
+
+! Not Backwards compatable !
+
+- adapted function-name geojson_to_multi
+- adapted predictor so point_snapping returns also predictions
+- Bugfix for SNAP_ALL_SIDE with big relevant distance (100m) resulting in wrong result [#199]
+- prediction-strategy: BEST/ALL/ORIGINAL: check if all records are found in result when using in combination with 'ONLY_FULL-strategy' [#200]
+- GRBACtualLoader: support for all types of GRB collections (also lines and points)
+- Fix for a better prediction of snapped lines